// native/api/model/marketCreateDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * SmartPreço API
 * SmartPreço API é uma API RESTful que fornece informações sobre mercados, produtos e seus preços.
 * OpenAPI spec version: 1.0
 */

export interface MarketCreateDto {
  /** Market's name */
  name: string;
  /** Market's address */
  address: string;
  /** Market's city */
  city: string;
  /** Market's state */
  state: string;
}