// native/api/model/productCreateDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * SmartPreço API
 * SmartPreço API é uma API RESTful que fornece informações sobre mercados, produtos e seus preços.
 * OpenAPI spec version: 1.0
 */

export interface ProductCreateDto {
  /** Product's name */
  name: string;
  /** Product's description */
  description: string;
  /** Product's category */
  category: string;
}